combine Area and NoiseArea in one class
